WebGraphics v0.0.1 WebGraphics.SVG.ElementBuilder

This module exposes the API for declaring

Functions

attr(elm, key, val)

Adds the given key and val to the elm map. The key is checked for possible handlers and aliases.

A map is returned with the modified element including the new attribute.

class(elm, val)

Adds the given val to the class property in the elm.

Example

rect(x: 5, y: 5) |> class("my-class my-other-class")

id(elm, val)

Adds the given val to the id property in the elm.

Example

rect(x: 10, y: 10) |> id("my-id")

render(elm)

Renders the given elm to a valid SVG string.

render_attr(key, val)

render_attrs(elm)

render_values(values)

render_with(elm, body)

style(elm, styles)

Alternative to style, so we can pass a map of properties to the element.

Example

rect(x: 10, y: 10) |> style(%{"fill" => "blue", "stroke" => "black"})

style(elm, key, val)

Adds the given key and val to the elm’s style property. This function can be called multiple times with different style properties.

Example

rect(x: 10, y: 10) |> style("fill", "red") |> style("stroke", "blue")