WebGraphics v0.0.1 WebGraphics.SVG.G

Functions

attr_alias(key)

Default implementation for the attr_alias function. Attribute aliasing works by defining multiple functions with pattern matching, and this function is the “match all” that just returns the given key.

attr_handler(, val)

Default implementation for the attr_handler function. Attribute handling works by defining multiple functions with pattern matching, and this function is the “match all” that just returns the given value.

elm(attrs \\ %{})

Returns a simple map with the given attrs and :_module that is a reference to the module that this function was called. This is necessary because each module defines different attribute aliases and handlers, and we need a way to get call those.

It’s possible to build up from the returned map using the functions in the WebGraphics.SVG.Element module, such as attr and style. After you’re done setting up the element

get_tag()

Returns the tag specified by the user. The tag must be a valid SVG element such as rect, circle or path. This function is called by the rener function in the ElementBuilder module.